# openapi.yaml
openapi: 3.0.3
info:
  title: Lumière E-Commerce API
  description: >
    OpenAPI specification for the Lumière e-commerce store (Next.js App Router API).
    Enables AI discovery, search, and recommendations via Navi AI Commerce.

    Authentication: protected endpoints accept either a NextAuth session cookie
    (next-auth.session-token) or a bearer JWT obtained from POST /api/auth/login
    (JSON, returns a signed JWT with expiry). Cookie login is performed via
    POST /api/auth/callback/credentials (form-encoded, no JSON token is returned).
    Admin management (users, orders, coupons, inventory) is implemented server-side
    and has no REST endpoints.
  version: 1.0.0
  license:
    name: Proprietary
servers:
  - url: https://lumier-commerce.vercel.app
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
        code:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    Role:
      type: string
      enum:
        - USER
        - ADMIN
        - SELLER
...
    AuthUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        avatar:
          type: string
          nullable: true
        emailVerified:
          type: string
          format: date-time
          nullable: true
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/AuthUser"
...
    ReviewUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
    ReviewReply:
      type: object
      properties:
        id:
          type: string
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/ReviewUser"
    Review:
      type: object
      properties:
        id:
          type: string
        rating:
          type: number
        comment:
          type: string
        isVerifiedPurchase:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/ReviewUser"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/ReviewReply"
...
    ReviewCreatedResponse:
      type: object
      properties:
        message:
          type: string
        review:
          $ref: "#/components/schemas/Review"
    ReplyCreatedResponse:
      type: object
      properties:
        message:
          type: string
        reply:
          $ref: "#/components/schemas/ReviewReply"
...
    CheckoutAddressInput:
      type: object
      required:
        - fullName
        - phone
        - country
        - city
        - street
      properties:
        fullName:
          type: string
        phone:
          type: string
        country:
          type: string
        city:
          type: string
        street:
          type: string
        postalCode:
          type: string
          nullable: true
    CheckoutInput:
      type: object
      required:
        - provider
        - shippingAddress
      properties:
        provider:
          type: string
          enum:
            - stripe
            - paypal
            - cod
        shippingAddress:
          $ref: "#/components/schemas/CheckoutAddressInput"
        promoCode:
          type: string
          nullable: true
        guestEmail:
          type: string
          nullable: true
    CheckoutStripe:
      type: object
      properties:
        clientSecret:
          type: string
    CheckoutPaypal:
      type: object
      properties:
        orderId:
          type: string
    CheckoutResult:
      type: object
      properties:
        orderId:
          type: string
        provider:
          type: string
          enum:
            - stripe
            - paypal
            - cod
        status:
          type: string
          enum:
            - pending
        amount:
          type: number
        currency:
          type: string
        stripe:
          $ref: "#/components/schemas/CheckoutStripe"
        paypal:
          $ref: "#/components/schemas/CheckoutPaypal"
...
    SellerProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        shopName:
          type: string
        description:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SellerRegisterResponse:
      type: object
      properties:
        message:
          type: string
        profile:
          $ref: "#/components/schemas/SellerProfile"
...
paths:
  /api/auth/login:
    post:
      summary: Log in with email and password, receive a bearer JWT
      description: >
        JSON POST. On success returns a signed HS256 JWT (expires in 1 day, or 30 days
        with rememberMe: true) plus the token expiry timestamp and the user record.
        Use the token as Authorization: Bearer <token> for protected endpoints.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
                rememberMe:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Login successful, JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Email not verified, account deleted, or account restricted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/checkout:
    post:
      summary: Start a checkout for the current cart
      description: Requires an authenticated session or a guest cart cookie (guestEmail required for guest checkout).
      tags:
        - Checkout
      security:
        - CookieAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "201":
          description: Checkout initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
        "400":
          description: Invalid checkout payload or missing guest email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required or missing guest cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/reviews:
    post:
      summary: Create a review for a purchased product
      tags:
        - Reviews
      security:
        - CookieAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - rating
                - comment
              properties:
                productId:
                  type: string
                rating:
                  type: number
                  minimum: 0
                  maximum: 5
                comment:
                  type: string
                  maxLength: 1000
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewCreatedResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not allowed to review this product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Review already exists for this product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/reviews/{id}:
    patch:
      summary: Update your own review
      tags:
        - Reviews
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: number
                  minimum: 0
                  maximum: 5
                comment:
                  type: string
                  maxLength: 1000
      responses:
        "200":
          description: Review updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewCreatedResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: You can only edit your own review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete your own review
      tags:
        - Reviews
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Review deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: You can only delete your own review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/reviews/{id}/replies:
    post:
      summary: Reply to a review
      tags:
        - Reviews
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
              properties:
                comment:
                  type: string
                  maxLength: 1000
      responses:
        "201":
          description: Reply added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyCreatedResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Review not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/review-replies/{id}:
    patch:
      summary: Update your own reply
      tags:
        - Reviews
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
              properties:
                comment:
                  type: string
                  maxLength: 1000
      responses:
        "200":
          description: Reply updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyCreatedResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: You can only edit your own reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Reply not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete your own reply
      tags:
        - Reviews
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reply deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: You can only delete your own reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Reply not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/seller/register:
    post:
      summary: Register an authenticated user as a seller
      tags:
        - Seller
      security:
        - CookieAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - shopName
              properties:
                shopName:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Seller registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerRegisterResponse"
        "400":
          description: Shop name is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Shop name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...